@@ -3,7 +3,7 @@ description: A new Flutter project.
 
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+publish_to: "none" # Remove this line if you wish to publish to pub.dev
 
 # The following defines the version and build number for your application.
 # A version number is three numbers separated by dots, like 1.2.43
@@ -20,9 +20,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.17.3 <3.0.0"
-  
-  
+  sdk: ">=3.4.3 <4.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -36,23 +34,14 @@ dependencies:
   flutter_model:
     path: ../../flutter_model
 
-  woue_components:
-    #path: ../../woue_components
-      git:
-       url: https://github.com/cracklye/woue_components.git
   bloc:
   flutter_bloc:
-  
+
   fluro: ^2.0.4
   fluent_ui:
   path:
   json_annotation:
   freezed: 2.3.2
-  woue_components_material: 
-    path: ../../woue_components/woue_components_material/
-
-  woue_components_fluent: 
-    path: ../../woue_components/woue_components_fluent/    
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
@@ -65,10 +54,8 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  build_runner: 
+  build_runner:
   json_serializable: ^6.2.0
-  
-  
 
   # The "flutter_lints" package below contains a set of recommended lints to
   # encourage good coding practices. The lint set provided by the package is
@@ -82,7 +69,6 @@ dev_dependencies:
 
 # The following section is specific to Flutter packages.
 flutter:
-
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
   # the material Icons class.
