@@ -1,17 +1,17 @@
 import 'package:example/notes/model_notes.dart';
 import 'package:example/notes/widget/note_display.dart';
 import 'package:example/notes/widget/note_form.dart';
-import 'package:example/repos/notes_repo_inmemory.dart';
 import 'package:example/repos/notes_repo_inmemory_slow.dart';
 import 'package:example/routes/app_navigator.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:flutter_model/bloc/modeledit/model_edit_state.dart';
 import 'package:flutter_model/flutter_model.dart';
+import 'package:flutter_model/screen/model_screen_list2.dart';
 
-class NotesScreenListSingleSlow extends ModelScreenList<Notes> {
-  NotesScreenListSingleSlow({Key? key})
+class NotesScreenListSingleSlow extends ModelScreenListActionBar<Notes> {
+  NotesScreenListSingleSlow({super.key})
       : super(
-          key: key,
           drawer: const AppDrawer(),
           title: "Notes",
         );
@@ -49,7 +49,7 @@ class NotesScreenListSingleSlow extends ModelScreenList<Notes> {
 
   @override
   Widget buildDetailDisplayForModel(
-      BuildContext context, ModelEditViewStateView<Notes> state, Notes? model) {
+      BuildContext context, ModelEditViewStateLoaded<Notes> state, Notes? model) {
     if (model == null) return const Text("Model is null");
     return NotesDisplay(model: model);
   }
@@ -57,7 +57,7 @@ class NotesScreenListSingleSlow extends ModelScreenList<Notes> {
   @override
   Widget buildForm(
     BuildContext context,
-    ModelEditViewStateEdit<Notes> state,
+    ModelEditViewStateLoaded<Notes> state,
     Notes? model,
     GlobalKey<FormState> formKey,
     void Function(Map<String, dynamic>) onSave,
