@@ -1,6 +1,6 @@
 import 'package:example/notes/model_notes.dart';
-import 'package:flutter/widgets.dart';
-import 'package:woue_components/woue_components.dart';
+import 'package:flutter/material.dart';
+
 
 class NotesListItem extends StatelessWidget {
   final Notes note;
