@@ -1,10 +1,10 @@
 
-import 'package:flutter_model/flutter_model.dart';
+// import 'package:flutter_model/flutter_model.dart';
 
 
-class ModelFilterEventFilter {
-  final List<Filter>? filters;
-  final List<SortOrderBy>? orderBy;
-  ModelFilterEventFilter({this.filters, this.orderBy});
-}
+// class ModelFilterEventFilter {
+//   final List<Filter>? filters;
+//   final List<SortOrderBy>? orderBy;
+//   ModelFilterEventFilter({this.filters, this.orderBy});
+// }
 
