@@ -1,35 +1,35 @@
 
-import 'package:flutter_model/flutter_model.dart';
+// import 'package:flutter_model/flutter_model.dart';
 
-abstract class ModelsDetailEvent<T extends IModel> {
-  const ModelsDetailEvent();
-}
+// abstract class ModelsDetailEvent<T extends IModel> {
+//   const ModelsDetailEvent();
+// }
 
-/// Raised when the selection criteria is changed or when the models need to
-/// be loaded for the first time.
-class ModelsDetailLoad<T extends IModel> extends ModelsDetailEvent<T> {
-  /// The parent ID that has been set to filter by
-  final dynamic id;
+// /// Raised when the selection criteria is changed or when the models need to
+// /// be loaded for the first time.
+// class ModelsDetailLoad<T extends IModel> extends ModelsDetailEvent<T> {
+//   /// The parent ID that has been set to filter by
+//   final dynamic id;
 
-  const ModelsDetailLoad({this.id});
-}
+//   const ModelsDetailLoad({this.id});
+// }
 
-/// Raised when a document needs to be deleted
-class ModelsDetailDelete<T extends IModel> extends ModelsDetailEvent<T> {
-  /// The model to delete
-  final dynamic id;
+// /// Raised when a document needs to be deleted
+// class ModelsDetailDelete<T extends IModel> extends ModelsDetailEvent<T> {
+//   /// The model to delete
+//   final dynamic id;
 
-  const ModelsDetailDelete(this.id);
-}
+//   const ModelsDetailDelete(this.id);
+// }
 
-class ModelsDetailUpdateDetail<T extends IModel> extends ModelsDetailEvent<T> {
-  final T model;
-  const ModelsDetailUpdateDetail(this.model);
-}
+// class ModelsDetailUpdateDetail<T extends IModel> extends ModelsDetailEvent<T> {
+//   final T model;
+//   const ModelsDetailUpdateDetail(this.model);
+// }
 
 
 
-class ModelDetailRaiseError<T extends IModel> extends ModelsDetailEvent<T> {
-  final String message;
-  const ModelDetailRaiseError(this.message);
-}
+// class ModelDetailRaiseError<T extends IModel> extends ModelsDetailEvent<T> {
+//   final String message;
+//   const ModelDetailRaiseError(this.message);
+// }
