@@ -6,17 +6,17 @@ import 'package:flutter_model/flutter_model.dart';
 
 mixin ModelListBlocAddin<T extends IModel> {
   Widget buildListBlocProvider(BuildContext context, dynamic parentId) {
-    return BlocProvider<ModelsListBloc<T>>(
-        create: ((context) => ModelsListBloc<T>(
+    return BlocProvider<ModelsBloc<T>>(
+        create: ((context) => ModelsBloc<T>(
             modelsRepository: RepositoryProvider.of<IModelAPI<T>>(context))
-          ..add(ModelsListLoad<T>(
+          ..add(ModelsLoad<T>(
             parentId: parentId,
           ))),
         child: buildListBlocBuilder(context));
   }
 
   Widget buildListBlocBuilder(BuildContext context) {
-    return BlocBuilder<ModelsListBloc<T>, ModelsListState<T>>(
+    return BlocBuilder<ModelsBloc<T>, ModelsState<T>>(
       builder: (context, state) {
         //  return buildBlocLoaded(context, state);
         return buildListBlocContent(context, state);
@@ -24,57 +24,57 @@ mixin ModelListBlocAddin<T extends IModel> {
     );
   }
 
-  Widget buildListBlocContent(BuildContext context, ModelsListState<T> state) {
-    if (state is ModelsListLoaded<T>) {
+  Widget buildListBlocContent(BuildContext context, ModelsState<T> state) {
+    if (state is ModelsLoaded<T>) {
       return buildListBlocLoaded(context, state);
-    } else if (state is ModelsListLoading<T>) {
+    } else if (state is ModelsLoading<T>) {
       return buildListBlocLoading(context, state);
     }
-    if (state is ModelsListNotLoaded<T>) {
+    if (state is ModelsNotLoaded<T>) {
       return buildListBlocNotLoaded(context, state);
     }
     return Text("Unknown state: $state");
   }
 
-  Widget buildListBlocLoaded(BuildContext context, ModelsListLoaded<T> state) {
+  Widget buildListBlocLoaded(BuildContext context, ModelsLoaded<T> state) {
     return Container();
   }
 
   Widget buildListBlocLoading(
-      BuildContext context, ModelsListLoading<T> state) {
+      BuildContext context, ModelsLoading<T> state) {
     return const Text("Loading");
   }
 
   Widget buildListBlocNotLoaded(
-      BuildContext context, ModelsListNotLoaded<T> state) {
+      BuildContext context, ModelsNotLoaded<T> state) {
     return const Text("Not Loaded");
   }
 
   void doListChangeSearch(BuildContext context, String searchString) {
-    BlocProvider.of<ModelsListBloc<T>>(context)
-        .add(ModelsListChangeSearchText<T>(searchString));
+    BlocProvider.of<ModelsBloc<T>>(context)
+        .add(ModelsChangeSearchText<T>(searchString));
   }
 
   // void doListChangeOrderBy(
   //     BuildContext context, SortOrderBy? orderByChange) {
-  //   BlocProvider.of<ModelsListBloc<T>>(context).add(ModelsListChangeOrderBy<T>(
+  //   BlocProvider.of<ModelsBloc<T>>(context).add(ModelsChangeOrderBy<T>(
   //       orderByChange != null ? orderByChange!.getSortOrders() : null));
   // }
 
   // void doListChangeFilter(BuildContext context, FilterByItem<T>? filter) {
-  //   BlocProvider.of<ModelsListBloc<T>>(context).add(
-  //       ModelsListChangeFilter<T>(filter != null ? filter.getFilters() : null));
+  //   BlocProvider.of<ModelsBloc<T>>(context).add(
+  //       ModelsChangeFilter<T>(filter != null ? filter.getFilters() : null));
   // }
 
-  bool isListLoaded(ModelsListState<T> state) {
-    if (state is ModelsListLoaded<T>) {
+  bool isListLoaded(ModelsState<T> state) {
+    if (state is ModelsLoaded<T>) {
       return true;
     }
     return false;
   }
 
-  bool isListLoading(ModelsListState<T> state) {
-    if (state is ModelsListLoading<T>) {
+  bool isListLoading(ModelsState<T> state) {
+    if (state is ModelsLoading<T>) {
       return true;
     }
     return false;
